refactor and add build target
