+++ Elements.CodeGeneration/src/Elements.CodeGeneration.targets
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+    <ItemGroup>
+        <TemplateFiles Include="$(MSBuildThisFileDirectory)\..\contentFiles\Templates\*.*" />
+    </ItemGroup>
+    <Target Name="CopyTemplateFiles" BeforeTargets="Build">
+        <Copy SourceFiles="@(TemplateFiles)" DestinationFolder="$(TargetDir)Templates\" />
+    </Target>
+</Project>
