+++ Elements.CodeGeneration/src/TypeGenerator.cs
@@ -3,9 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using DotLiquid;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Newtonsoft.Json;
@@ -14,104 +12,11 @@ using NJsonSchema.CodeGeneration.CSharp;
 using Elements.Generate.StringUtils;
 using NJsonSchema.CodeGeneration;
 using NJsonSchema.CodeGeneration.CSharp.Models;
+using Elements.Geometry;
 
-namespace Elements.Generate.StringUtils
-{
-    public static class StringUtilities
-    {
-        // TODO: This is a direct copy of the method in Hypar.Model.
-        // We should move type generation out of elements to a place where it can refer to Hypar.Model.
-        public static string ToSafeIdentifier(this string name, bool firstCharLowercase = false)
-        {
-            if (String.IsNullOrWhiteSpace(name) || String.IsNullOrEmpty(name))
-            {
-                throw new ArgumentException("Invalid name.");
-            }
-            //remove special characters except for space, dash, period, and underscore
-            name = Regex.Replace(name, @"[^A-Za-z0-9 _\.-]", "");
-            //remove any numbers or whitespace/separator characters from the front
-            while (Char.IsDigit(name.First()) || name.First() == ' ' || name.First() == '_' || name.First() == '-' || name.First() == '.')
-            {
-                if (name.Length < 2)
-                {
-                    throw new ArgumentException("Name is invalid: nothing is left after removing numeric characters. Names must contain at least one letter.");
-                }
-                name = name.Substring(1);
-            }
-            //split on whitespace or - or _
-            var splits = name.Split(new[] { ' ', '-', '_', '.' });
-            var cleanName = "";
-            var dontCapitalize = firstCharLowercase;
-            foreach (var split in splits)
-            {
-                if (split.Length == 0)
-                {
-                    continue;
-                }
-                if (dontCapitalize)
-                {
-                    cleanName += split.First().ToString().ToLower();
-                    dontCapitalize = false;
-                }
-                else
-                {
-                    cleanName += split.First().ToString().ToUpper();
-                }
-                if (split.Length > 1)
-                {
-                    cleanName += split.Substring(1);
-                }
-            }
-            if (string.IsNullOrEmpty(cleanName))
-            {
-                throw new ArgumentException("Names must have at least one letter character.");
-            }
-            return cleanName;
-        }
-
-    }
-
-}
 
 namespace Elements.Generate
 {
-    /// <summary>
-    /// The result of a compilation.
-    /// </summary>
-    public struct CompilationResult
-    {
-        /// <summary>
-        /// True if the compilation succeeded.
-        /// </summary>
-        public bool Success { get; internal set; }
-        /// <summary>
-        /// The Assembly loaded from the compilation, if successful.
-        /// </summary>
-        public Assembly Assembly { get; internal set; }
-        /// <summary>
-        /// Any messages or errors that arose during compilation.
-        /// </summary>
-        public string[] DiagnosticResults { get; internal set; }
-    }
-
-    /// <summary>
-    /// The result of code generation.
-    /// </summary>
-    public struct GenerationResult
-    {
-        /// <summary>
-        /// True if the code was generated successfully.
-        /// </summary>
-        public bool Success { get; internal set; }
-        /// <summary>
-        /// The file path to the generated code.
-        /// </summary>
-        public string FilePath { get; internal set; }
-        /// <summary>
-        /// Any messages or errors that arose during code generation.
-        /// </summary>
-        public string[] DiagnosticResults { get; internal set; }
-    }
     class ElementsTypeNameGenerator : ITypeNameGenerator
     {
         // TODO(Ian): This type name generator is only required because njsonschema
@@ -142,6 +47,7 @@ namespace Elements.Generate
             return property.Name.ToSafeIdentifier();
         }
     }
+
     /// <summary>
     /// TypeGenerator contains logic for generating element types from JSON schemas.
     /// </summary>
@@ -193,47 +99,16 @@ namespace Elements.Generate
                 {
                     _templatesPath = Path.GetFullPath(Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "./Templates"));
                 }
+                if (!Directory.Exists(_templatesPath))
+                {
+                    Console.WriteLine("Templates path attempted: " + _templatesPath);
+                    throw new InvalidDataException("The templates folder cannot be found and is necessary for successful code generation. Be sure that the Hypar.Elements.CodeGeneration nuget package is referenced in all of your projects.");
+                }
                 return _templatesPath;
             }
             set => _templatesPath = value;
         }
 
-        // TODO Delete this HyparFilters class when this issue gets resolved. https://github.com/RicoSuter/NJsonSchema/issues/1199
-        // This HyparFilters class contains filters that are copied directly from the NJsonSchema repo
-        // because the filters are not public but we need to register them globally for async code gen.
-        // Copied from https://github.com/RicoSuter/NJsonSchema/blob/687efeabdc30ddacd235e85213f3594458ed48b4/src/NJsonSchema.CodeGeneration/DefaultTemplateFactory.cs#L183
-        public static class HyparFilters
-        {
-            public static string Lowercamelcase(Context context, string input, bool firstCharacterMustBeAlpha = true)
-            {
-                return ConversionUtilities.ConvertToLowerCamelCase(input, firstCharacterMustBeAlpha);
-            }
-
-            public static string Safeidentifierlower(Context context, string input, bool firstCharacterMustBeAlpha = true)
-            {
-                return input.ToSafeIdentifier(true);
-            }
-
-            public static string Safeidentifierupper(Context context, string input, bool firstCharacterMustBeAlpha = true)
-            {
-                return input.ToSafeIdentifier();
-            }
-
-            public static string Csharpdocs(string input, int tabCount)
-            {
-                return ConversionUtilities.ConvertCSharpDocs(input, tabCount);
-            }
-
-            public static IEnumerable<object> Empty(Context context, object input)
-            {
-                return Enumerable.Empty<object>();
-            }
-
-            public static string Tab(Context context, string input, int tabCount)
-            {
-                return ConversionUtilities.Tab(input, tabCount);
-            }
-        }
 
         /// <summary>
         /// Generate a user-defined type in a .g.cs file from a schema.
@@ -750,7 +625,6 @@ using Hypar.Functions.Execution.AWS;", "");
             {
                 var tree = CSharpSyntaxTree.ParseText(cs, options);
                 syntaxTrees.Add(tree);
-
             }
 
             var assemblyPath = frameworkBuild ? @"C:\Windows\Microsoft.NET\Framework64\v4.0.30319" : Path.GetDirectoryName(typeof(object).Assembly.Location);
@@ -758,7 +632,7 @@ using Hypar.Functions.Execution.AWS;", "");
             var newtonSoftPath = Path.GetDirectoryName(typeof(JsonConverter).Assembly.Location);
 
             IEnumerable<MetadataReference> defaultReferences = new[]
-           {
+            {
                 MetadataReference.CreateFromFile(Path.Combine(assemblyPath, "netstandard.dll")),
                 MetadataReference.CreateFromFile(Path.Combine(assemblyPath, "System.ComponentModel.Annotations.dll")),
                 MetadataReference.CreateFromFile(Path.Combine(assemblyPath, "System.Diagnostics.Tools.dll")),
