+++ Elements/src/Elements.csproj
@@ -1,37 +1,37 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>Hypar.Elements</AssemblyName>
-    <PackageTitle>Hypar Elements</PackageTitle>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <Title>Hypar Elements</Title>
-    <PackageId>Hypar.Elements</PackageId>
-    <PackageDescription>A building elements library for AEC.</PackageDescription>
-    <Summary>The Elements library provides object types for generating the built environment.</Summary>
-    <PackageProjectUrl>https://github.com/hypar-io/elements</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/hypar-io/elements</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <Version>$(Version)</Version>
-    <PackageOutputPath>../../nupkg</PackageOutputPath>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>netstandard2.0</TargetFramework>
+        <AssemblyName>Hypar.Elements</AssemblyName>
+        <PackageTitle>Hypar Elements</PackageTitle>
+        <GenerateDocumentationFile>true</GenerateDocumentationFile>
+        <Title>Hypar Elements</Title>
+        <PackageId>Hypar.Elements</PackageId>
+        <PackageDescription>A building elements library for AEC.</PackageDescription>
+        <Summary>The Elements library provides object types for generating the built environment.</Summary>
+        <PackageProjectUrl>https://github.com/hypar-io/elements</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/hypar-io/elements</RepositoryUrl>
+        <RepositoryType>git</RepositoryType>
+        <Version>$(Version)</Version>
+        <PackageOutputPath>../../nupkg</PackageOutputPath>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="glTF2Loader" Version="1.1.3-alpha" />
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-    <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0" />
-    <PackageReference Include="System.ComponentModel.Annotations" Version="4.6.0" />
-    <PackageReference Include="Unofficial.LibTessDotNet" Version="2.0.0" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="glTF2Loader" Version="1.1.3-alpha" />
+        <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
+        <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0" />
+        <PackageReference Include="System.ComponentModel.Annotations" Version="4.6.0" />
+        <PackageReference Include="Unofficial.LibTessDotNet" Version="2.0.0" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <Content Include="Textures\**\*.*">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>   
-  </ItemGroup>
-  
-  <ItemGroup>
-    <None Update="UV.jpg" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
+    <ItemGroup>
+        <Content Include="Textures\**\*.*">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </Content>
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Update="UV.jpg" CopyToOutputDirectory="PreserveNewest" />
+    </ItemGroup>
 
 </Project>
